Add Linux master user rpmsg driver and user apps

+++ proxy/master/linux/userspace/proxy_app.h
@@ -20,13 +20,13 @@ struct _rpc_data {
 struct _sys_call_args {
 	int int_field1;
 	int int_field2;
-	unsigned long	data_len;
+	unsigned int	data_len;
 	char data[0];
 };
 
 /* System call rpc data structure */
 struct _sys_rpc {
-	unsigned long	id;
+	unsigned int	id;
 	struct _sys_call_args	sys_call_args;
 };
